+++ quick-order/src/main/java/com/quick_order/entity/TableInfo.java
@@ -19,9 +19,9 @@ public class TableInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String tableNo;
-    private Integer outletId;
+    private Long outletId;
     private Integer sittingCapacity;
 
 }
